change interface and set docs in function files for simple way to find documentation for function in code
