+++ src/Documentation/SimpleDocumentation.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <memory>
+#include <Documentation/IDocumentation.h>
+
+namespace DB
+{
+
+class SimpleDocumentation : public IDocumentation
+{
+public:
+    explicit SimpleDocumentation(const char* docs);
+
+    std::string getDocumentation() const override;
+};
+
+IDocumentationPtr makeSimpleDocumentation(const char * docs);
+
+}
