--- src/Documentation/SimpleDocumentationMerge.h
@@ -1,27 +1,0 @@
-#pragma once
-
-#include <Documentation/IDocumentation.h>
-
-namespace DB
-{
-
-namespace MergeDoc
-{
-const char * doc = R"(
----
-toc_priority: 38
-toc_title: merge
----
-
-# merge {#merge}
-
-`merge(db_name, 'tables_regexp')` – Creates a temporary Merge table. For more information, see the section “Table engines, Merge”.
-
-The table structure is taken from the first table encountered that matches the regular expression.
-
-
-)";
-
-}
-
-}
