--- src/Documentation/SimpleDocumentationOdbc.h
@@ -1,120 +1,0 @@
-#pragma once
-
-#include <Documentation/IDocumentation.h>
-
-namespace DB
-{
-
-namespace OdbcDoc
-{
-const char * doc = R"(
----
-toc_priority: 44
-toc_title: odbc
----
-
-# odbc {#table-functions-odbc}
-
-Returns table that is connected via [ODBC](https://en.wikipedia.org/wiki/Open_Database_Connectivity).
-
-``` sql
-odbc(connection_settings, external_database, external_table)
-```
-
-Parameters:
-
--   `connection_settings` — Name of the section with connection settings in the `odbc.ini` file.
--   `external_database` — Name of a database in an external DBMS.
--   `external_table` — Name of a table in the `external_database`.
-
-To safely implement ODBC connections, ClickHouse uses a separate program `clickhouse-odbc-bridge`. If the ODBC driver is loaded directly from `clickhouse-server`, driver problems can crash the ClickHouse server. ClickHouse automatically starts `clickhouse-odbc-bridge` when it is required. The ODBC bridge program is installed from the same package as the `clickhouse-server`.
-
-The fields with the `NULL` values from the external table are converted into the default values for the base data type. For example, if a remote MySQL table field has the `INT NULL` type it is converted to 0 (the default value for ClickHouse `Int32` data type).
-
-## Usage Example {#usage-example}
-
-**Getting data from the local MySQL installation via ODBC**
-
-This example is checked for Ubuntu Linux 18.04 and MySQL server 5.7.
-
-Ensure that unixODBC and MySQL Connector are installed.
-
-By default (if installed from packages), ClickHouse starts as user `clickhouse`. Thus you need to create and configure this user in the MySQL server.
-
-``` bash
-$ sudo mysql
-```
-
-``` sql
-mysql> CREATE USER 'clickhouse'@'localhost' IDENTIFIED BY 'clickhouse';
-mysql> GRANT ALL PRIVILEGES ON *.* TO 'clickhouse'@'clickhouse' WITH GRANT OPTION;
-```
-
-Then configure the connection in `/etc/odbc.ini`.
-
-``` bash
-$ cat /etc/odbc.ini
-[mysqlconn]
-DRIVER = /usr/local/lib/libmyodbc5w.so
-SERVER = 127.0.0.1
-PORT = 3306
-DATABASE = test
-USERNAME = clickhouse
-PASSWORD = clickhouse
-```
-
-You can check the connection using the `isql` utility from the unixODBC installation.
-
-``` bash
-$ isql -v mysqlconn
-+-------------------------+
-| Connected!                            |
-|                                       |
-...
-```
-
-Table in MySQL:
-
-``` text
-mysql> CREATE TABLE `test`.`test` (
-    ->   `int_id` INT NOT NULL AUTO_INCREMENT,
-    ->   `int_nullable` INT NULL DEFAULT NULL,
-    ->   `float` FLOAT NOT NULL,
-    ->   `float_nullable` FLOAT NULL DEFAULT NULL,
-    ->   PRIMARY KEY (`int_id`));
-Query OK, 0 rows affected (0,09 sec)
-
-mysql> insert into test (`int_id`, `float`) VALUES (1,2);
-Query OK, 1 row affected (0,00 sec)
-
-mysql> select * from test;
-+------+----------+-----+----------+
-| int_id | int_nullable | float | float_nullable |
-+------+----------+-----+----------+
-|      1 |         NULL |     2 |           NULL |
-+------+----------+-----+----------+
-1 row in set (0,00 sec)
-```
-
-Retrieving data from the MySQL table in ClickHouse:
-
-``` sql
-SELECT * FROM odbc('DSN=mysqlconn', 'test', 'test')
-```
-
-``` text
-┌─int_id─┬─int_nullable─┬─float─┬─float_nullable─┐
-│      1 │            0 │     2 │              0 │
-└────────┴──────────────┴───────┴────────────────┘
-```
-
-## See Also {#see-also}
-
--   [ODBC external dictionaries](../../sql-reference/dictionaries/external-dictionaries/external-dicts-dict-sources.md#dicts-external_dicts_dict_sources-odbc)
--   [ODBC table engine](../../engines/table-engines/integrations/odbc.md).
-
-)";
-
-}
-
-}
