+++ src/TableFunctions/TableFunctionFactory.cpp
@@ -18,19 +18,18 @@ namespace ErrorCodes
 }
 
 
-void TableFunctionFactory::registerFunction(const std::string & name, Value creator, CaseSensitiveness case_sensitiveness, const char* documentation)
+void TableFunctionFactory::registerFunction(const std::string & name, Value creator, CaseSensitiveness case_sensitiveness, IDocumentationPtr documentation)
 {
-    IDocumentation* docs = new IDocumentation(documentation);
     if (!table_functions.emplace(name, creator).second)
         throw Exception("TableFunctionFactory: the table function name '" + name + "' is not unique",
             ErrorCodes::LOGICAL_ERROR);
-    else table_docs.emplace(name, docs);
+    else table_docs.emplace(name, std::move(documentation));
 
     if (case_sensitiveness == CaseInsensitive
         && !case_insensitive_table_functions.emplace(Poco::toLower(name), creator).second)
         throw Exception("TableFunctionFactory: the case insensitive table function name '" + name + "' is not unique",
                         ErrorCodes::LOGICAL_ERROR);
-    else case_insensitive_table_docs.emplace(name, docs);
+    else case_insensitive_table_docs.emplace(name, std::move(documentation));
 }
 
 TableFunctionPtr TableFunctionFactory::get(
@@ -82,18 +81,18 @@ TableFunctionPtr TableFunctionFactory::tryGet(
     return res;
 }
 
-const char* TableFunctionFactory::getDocumetation(const std::string & name_param) const
+std::string TableFunctionFactory::getDocumetation(const std::string & name_param) const
 {
     String name = getAliasToOrName(name_param);
 
     auto it = table_docs.find(name);
     if (table_docs.end() != it)
-        return it->second->getDocumentation();
+        return it->second == nullptr ? "Not found" : it->second->getDocumentation();
     else
     {
         it = case_insensitive_table_docs.find(Poco::toLower(name));
         if (case_insensitive_table_docs.end() != it)
-            return it->second->getDocumentation();
+            return it->second == nullptr ? "Not found" : it->second->getDocumentation();
     }
 
     if (CurrentThread::isInitialized())
