+++ src/TableFunctions/TableFunctionFactory.h
@@ -12,6 +12,7 @@
 #include <unordered_map>
 #include <boost/noncopyable.hpp>
 #include <Documentation/IDocumentation.h>
+#include <Documentation/SimpleDocumentation.h>
 
 
 namespace DB
@@ -30,16 +31,16 @@ public:
 
     /// Register a function by its name.
     /// No locking, you must register all functions before usage of get.
-    void registerFunction(const std::string & name, Value creator, CaseSensitiveness case_sensitiveness = CaseSensitive, const char* documentation="Not found");
+    void registerFunction(const std::string & name, Value creator, CaseSensitiveness case_sensitiveness = CaseSensitive, IDocumentationPtr documentation=nullptr);
 
     template <typename Function>
-    void registerFunction(CaseSensitiveness case_sensitiveness = CaseSensitive, const char* documentation="Not found")
+    void registerFunction(CaseSensitiveness case_sensitiveness = CaseSensitive, IDocumentationPtr documentation=nullptr)
     {
         auto creator = [] () -> TableFunctionPtr
         {
             return std::make_shared<Function>();
         };
-        registerFunction(Function::name, std::move(creator), case_sensitiveness, documentation);
+        registerFunction(Function::name, std::move(creator), case_sensitiveness, std::move(documentation));
     }
 
     /// Throws an exception if not found.
@@ -51,7 +52,7 @@ public:
     bool isTableFunctionName(const std::string & name) const;
 
     /// If there will be no documentation returns "Not found"
-    const char* getDocumetation(const std::string & name) const;
+    std::string getDocumetation(const std::string & name) const;
 private:
     using TableFunctions = std::unordered_map<std::string, Value>;
     using TableFunctionsDocs = std::unordered_map<std::string, IDocumentationPtr>;
