+++ src/TableFunctions/TableFunctionHDFS.cpp
@@ -6,7 +6,6 @@
 #include <Storages/ColumnsDescription.h>
 #include <TableFunctions/TableFunctionFactory.h>
 #include <TableFunctions/TableFunctionHDFS.h>
-#include <Documentation/SimpleDocumentationHdfs.h>
 
 namespace DB
 {
@@ -29,7 +28,7 @@ StoragePtr TableFunctionHDFS::getStorage(
 #if USE_HDFS
 void registerTableFunctionHDFS(TableFunctionFactory & factory)
 {
-    factory.registerFunction<TableFunctionHDFS>(TableFunctionFactory::CaseInsensitive, HdfsDoc::doc);
+    factory.registerFunction<TableFunctionHDFS>(TableFunctionFactory::CaseInsensitive, makeSimpleDocumentation(HdfsDoc::doc));
 }
 #endif
 }
