+++ src/TableFunctions/TableFunctionMerge.h
@@ -2,7 +2,6 @@
 
 #include <TableFunctions/ITableFunction.h>
 
-
 namespace DB
 {
 
@@ -33,5 +32,16 @@ private:
     mutable std::optional<DbToTableSetMap> source_databases_and_tables;
 };
 
+namespace MergeDoc
+{
+const char * doc = R"(
+`merge(db_name, 'tables_regexp')` – Creates a temporary Merge table. For more information, see the section “Table engines, Merge”.
+
+The table structure is taken from the first table encountered that matches the regular expression.
+
+
+)";
+
+}
 
 }
