+++ src/TableFunctions/TableFunctionMySQL.cpp
@@ -26,7 +26,6 @@
 
 #include <Databases/MySQL/DatabaseMySQL.h> // for fetchTablesColumnsList
 #include <Common/parseRemoteDescription.h>
-#include <Documentation/SimpleDocumentationMysql.h>
 
 
 namespace DB
@@ -121,7 +120,7 @@ StoragePtr TableFunctionMySQL::executeImpl(
 
 void registerTableFunctionMySQL(TableFunctionFactory & factory)
 {
-    factory.registerFunction<TableFunctionMySQL>(TableFunctionFactory::CaseInsensitive, MysqlDoc::doc);
+    factory.registerFunction<TableFunctionMySQL>(TableFunctionFactory::CaseInsensitive, makeSimpleDocumentation(MysqlDoc::doc));
 }
 }
