+++ src/TableFunctions/TableFunctionPostgreSQL.cpp
@@ -14,7 +14,6 @@
 #include "registerTableFunctions.h"
 #include <Common/quoteString.h>
 #include <Common/parseRemoteDescription.h>
-#include <Documentation/SimpleDocumentationPostgresql.h>
 
 
 namespace DB
@@ -97,7 +96,7 @@ void TableFunctionPostgreSQL::parseArguments(const ASTPtr & ast_function, Contex
 
 void registerTableFunctionPostgreSQL(TableFunctionFactory & factory)
 {
-    factory.registerFunction<TableFunctionPostgreSQL>(TableFunctionFactory::CaseInsensitive, PostgresqlDoc::doc);
+    factory.registerFunction<TableFunctionPostgreSQL>(TableFunctionFactory::CaseInsensitive, makeSimpleDocumentation(PostgresqlDoc::doc));
 }
 
 }
